refactor(mobile): split pairing onboarding widgets
